Use scale-type-resolver types
